improve: move user option into own component
