+++ frontend/src/components/Navbar.tsx
@@ -7,14 +7,13 @@ import Typography from "@mui/material/Typography";
 import Menu from "@mui/material/Menu";
 import MenuIcon from "@mui/icons-material/Menu";
 import Container from "@mui/material/Container";
-import Avatar from "@mui/material/Avatar";
 import Button from "@mui/material/Button";
-import Tooltip from "@mui/material/Tooltip";
 import MenuItem from "@mui/material/MenuItem";
 import MusicNoteIcon from "@mui/icons-material/MusicNote";
 import { Link } from "react-router-dom";
 import { navbarStyles } from "../styles";
 import { useAuth0 } from "@auth0/auth0-react";
+import UserOptions from "../pages/users/UserOptions";
 
 const pages = [
   { name: "Home", path: "/" },
@@ -22,33 +21,18 @@ const pages = [
   { name: "Note Game", path: "/note-game" },
   { name: "About", path: "/about" },
 ];
-const settings = [
-  { name: "Profile", path: "/profile" },
-  { name: "Account", path: "/account" },
-  { name: "Dashboard", path: "/dashboard" },
-];
 
 function NavBar() {
-  const { loginWithRedirect, logout, user, isAuthenticated } = useAuth0();
+  const { loginWithRedirect, isAuthenticated } = useAuth0();
   const [anchorElNav, setAnchorElNav] = React.useState<null | HTMLElement>(
-    null,
-  );
-  const [anchorElUser, setAnchorElUser] = React.useState<null | HTMLElement>(
     null,
   );
   const handleOpenNavMenu = (event: React.MouseEvent<HTMLElement>) => {
     setAnchorElNav(event.currentTarget);
   };
-  const handleOpenUserMenu = (event: React.MouseEvent<HTMLElement>) => {
-    setAnchorElUser(event.currentTarget);
-  };
 
   const handleCloseNavMenu = () => {
     setAnchorElNav(null);
-  };
-
-  const handleCloseUserMenu = () => {
-    setAnchorElUser(null);
   };
 
   return (
@@ -130,55 +114,13 @@ function NavBar() {
             ))}
           </Box>
           <Box sx={{ flexGrow: 0 }}>
-            <Tooltip title="Open settings">
-              <IconButton onClick={handleOpenUserMenu} sx={{ p: 0 }}>
-                {isAuthenticated ? (
-                  <Avatar alt="Remy Sharp" src={user?.picture} />
-                ) : (
-                  <Avatar alt="Remy Sharp" src="/static/images/avatar/2.jpg" />
-                )}
-              </IconButton>
-            </Tooltip>
-            <Menu
-              sx={{ mt: "45px" }}
-              id="menu-appbar"
-              anchorEl={anchorElUser}
-              anchorOrigin={{
-                vertical: "top",
-                horizontal: "right",
-              }}
-              keepMounted
-              transformOrigin={{
-                vertical: "top",
-                horizontal: "right",
-              }}
-              open={Boolean(anchorElUser)}
-              onClose={handleCloseUserMenu}
-            >
-              <MenuItem key="Log in" onClick={() => loginWithRedirect()}>
-                <Typography sx={{ textAlign: "center" }}>Log In</Typography>
-              </MenuItem>
-              {settings.map((setting) => (
-                <MenuItem
-                  key={setting.name}
-                  component={Link}
-                  to={setting.path}
-                  onClick={handleCloseUserMenu}
-                >
-                  <Typography sx={{ textAlign: "center" }}>
-                    {setting.name}
-                  </Typography>
-                </MenuItem>
-              ))}
-              <MenuItem
-                key="Log Out"
-                onClick={() =>
-                  logout({ logoutParams: { returnTo: window.location.origin } })
-                }
-              >
-                <Typography sx={{ textAlign: "center" }}>Log Out</Typography>
-              </MenuItem>
-            </Menu>
+            {isAuthenticated ? (
+              <UserOptions />
+            ) : (
+              <Button color="inherit" onClick={() => loginWithRedirect()}>
+                Sign In
+              </Button>
+            )}
           </Box>
         </Toolbar>
       </Container>
